+++ gaios/HWFlow/Scan/ScanListItem.swift
@@ -3,7 +3,7 @@ import Foundation
 struct ScanListItem {
     let identifier: UUID
     let name: String
-    var ledger: Bool { name.contains("Nano") }
-    var jade: Bool { name.contains("Jade") }
-    var type: DeviceType? { jade ? .Jade : ledger ? .Ledger : nil }
+    var ledger: Bool { type == .Ledger }
+    var jade: Bool { type == .Jade }
+    var type: DeviceType
 }