+++ gaios/Shared/Controllers/DialogSendHWSummaryViewController.swift
@@ -91,7 +91,7 @@ class DialogSendHWSummaryViewController: UIViewController {
         if let outputs = transaction.transactionOutputs, !outputs.isEmpty {
             var changeAddress = [String]()
             outputs.forEach { output in
-                if output.isChange, let address = output.address {
+                if output.isChange ?? false, let address = output.address {
                     changeAddress.append(address)
                 }
             }