+++ hw/Ledger/BleLedger.swift
@@ -4,7 +4,7 @@ import greenaddress
 public class BleLedger: BleLedgerCommands, HWProtocol {
     
     let SIGHASH_ALL: UInt8 = 1
-
+    
     public func xpubs(network: String, paths: [[Int]]) async throws -> [String] {
         var res = [String]()
         for path in paths {
@@ -12,7 +12,7 @@ public class BleLedger: BleLedgerCommands, HWProtocol {
         }
         return res
     }
-
+    
     public func xpubs(network: String, path: [Int]) async throws -> String {
         let isMainnet = ["mainnet", "liquid"].contains(network)
         let data = try await pubKey(path: path)
@@ -29,47 +29,84 @@ public class BleLedger: BleLedgerCommands, HWProtocol {
         let signature = res["signature"] as? [UInt8]
         return HWSignMessageResult(signature: signature?.hex, signerCommitment: nil)
     }
-
+    
     public func signTransaction(network: String, params: HWSignTxParams) async throws -> HWSignTxResponse {
+        if params.useAeProtocol {
+            throw HWError.Abort("Hardware Wallet does not support the Anti-Exfil protocol");
+        }
+        var sw = false
+        var p2sh = false
+        params.signingInputs.forEach {
+            if $0.isSegwit {
+                sw = true
+            } else {
+                p2sh = true
+            }
+        }
+        if p2sh || !sw {
+            throw HWError.Abort("Supported only segwit");
+        }
+        
         let res = try await signSW(tx: params.transaction,
                          inputs: params.signingInputs,
-                         outputs: params.txOutputs)
+                         outputs: params.txOutputs,
+                         transactions: params.signingTxs
+        )
         return HWSignTxResponse(signatures: res, signerCommitments: [])
     }
 
-    public func signSW(tx: HWTransaction?, inputs: [InputOutput], outputs: [InputOutput]) async throws -> [String] {
-        let hwInputs = inputs.map { input -> [String: Any] in
-            let bytes = inputBytes(input, isSegwit: true)!
-            let sequence = input.sequence.uint32LE()
-            return ["value": bytes, "sequence": Data(sequence), "trusted": false, "segwit": true]
+    public func signSW(tx: String?, inputs: [InputOutput], outputs: [InputOutput], transactions: [String: String]) async throws -> [String] {
+        var hwInputs = [[String: Any]]()
+        for input in inputs {
+            guard let txHex = transactions[input.txHash ?? ""] else {
+                throw HWError.Abort("Previous transaction \(input.txHash ?? "") not found")
+            }
+            let txData = txHex.hexToData()
+            let tx = HWTransactionLedger(data: txData)
+            let trustedInput = try await getTrustedInput(transaction: tx, index: Int(input.ptIdx), sequence: input.sequence, segwit: true)
+            hwInputs += [trustedInput]
         }
+        
         // Prepare the pseudo transaction
         // Provide the first script instead of a null script to initialize the P2SH confirmation logic
-        let prevoutScript = inputs.first?.prevoutScript?.hexToData()
-        _ = try await startUntrustedTransaction(txVersion: tx?.transactionVersion ?? 0, newTransaction: true, inputIndex: 0, usedInputList: hwInputs, redeemScript: prevoutScript!, segwit: true)
+        guard let wallyTx = Wally.txFromBytes(tx: tx?.hexToBytes() ?? [], elements: false, segwit: true) else {
+            throw HWError.Abort("Invalid transaction")
+        }
+        
+        let version = Wally.txVersion(wallyTx: wallyTx)
+        let script0 = inputs.first?.prevoutScript?.hexToData()
+        _ = try await startUntrustedTransaction(txVersion: version, newTransaction: true, inputIndex: 0, usedInputList: hwInputs, redeemScript: script0!, segwit: true)
         let bytes = outputBytes(outputs)
         _ = try await finalizeInputFull(data: bytes!)
-        let sigs = try await signSWInputs(hwInputs: hwInputs, inputs: inputs, version: tx?.transactionVersion ?? 0, locktime: tx?.transactionLocktime ?? 0)
+        let locktime = Wally.txLocktime(wallyTx: wallyTx)
+        
+        // Sign each input
+        var sigs = [Data]()
+        for hwInput in hwInputs.enumerated() {
+            let input = inputs[hwInput.offset]
+            if !input.isSegwit { continue } // sign segwit only
+            let singleInput = [hwInput.element]
+            let script = input.prevoutScript?.hexToData()
+            _ = try await startUntrustedTransaction(txVersion: version,
+                                                    newTransaction: false,
+                                                    inputIndex: 0,
+                                                    usedInputList: singleInput,
+                                                    redeemScript: script!,
+                                                    segwit: true)
+            let userPaths: [Int] = input.userPath?.map { Int($0) } ?? []
+            let sig = try await untrustedHashSign(privateKeyPath: userPaths,
+                                                  pin: "0",
+                                                  lockTime: locktime,
+                                                  sigHashType: self.SIGHASH_ALL)
+            sigs += [sig]
+        }
         var strings = [String]()
-        for sig in sigs {
-            strings += [sig.hex]
+        for input in inputs.enumerated() {
+            if input.offset < sigs.count {
+                strings += [sigs[input.offset].hex]
+            }
         }
         return strings
-    }
-
-    public func signSWInputs(hwInputs: [[String: Any]], inputs: [InputOutput], version: UInt, locktime: UInt) async throws -> [Data] {
-        var list = [Data]()
-        for hwInput in hwInputs.enumerated() {
-            list += [try await signSWInput(hwInput: hwInput.element, input: inputs[hwInput.offset], version: version, locktime: locktime)]
-        }
-        return list
-    }
-
-    public func signSWInput(hwInput: [String: Any], input: InputOutput, version: UInt, locktime: UInt) async throws -> Data {
-        let prevoutScript = input.prevoutScript!.hexToData()
-        _ = try await startUntrustedTransaction(txVersion: version, newTransaction: false, inputIndex: 0, usedInputList: [hwInput], redeemScript: prevoutScript, segwit: true)
-        let userPaths: [Int] = input.userPath?.map { Int($0) } ?? []
-        return try await untrustedHashSign(privateKeyPath: userPaths, pin: "0", lockTime: locktime, sigHashType: self.SIGHASH_ALL)
     }
 
     // swiftlint:disable:next function_parameter_count