+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Update gdk to 0.0.65
+- Improve signing on Ledger
 
 ## [4.0.10] - 2023-07-25